Replaced Old VideoCards.js and CompactVideoCard.js
